requirements.txt
fastf1
pandas
numpy
scikit-learn
streamlit
plotly
python-dotenv
requests
xgboost
